# openapi.yaml
openapi: 3.0.3
info:
  title: LLM Orchestra API
  description: |
    Natural Language Orchestrator API for Gmail, Google Calendar, and Google Drive.
    
    This API allows you to control Google services using plain English commands.
    It uses GPT-4 for intent parsing and coordinates across multiple services.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: Google OAuth authentication
  - name: Commands
    description: Natural language command processing
  - name: Session
    description: Session management
  - name: Gmail
    description: Gmail operations
  - name: Calendar
    description: Calendar operations
  - name: Drive
    description: Drive operations

paths:
  /:
    get:
      tags:
        - Health
      summary: Root endpoint
      description: Returns API information
      responses:
        '200':
          description: API information
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: LLM Orchestra API
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: string
                    example: running

  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check API health status
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/v1/auth:
    post:
      tags:
        - Authentication
      summary: Authenticate with Google
      description: Initiate Google OAuth authentication flow
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: authenticated
                  message:
                    type: string
                    example: Successfully authenticated with Google
        '401':
          description: Authentication failed
        '500':
          description: Server error

  /api/v1/command:
    post:
      tags:
        - Commands
      summary: Process natural language command
      description: |
        Process a natural language command and execute the appropriate action.
        
        Examples:
        - "search for emails from google"
        - "what's my next meeting?"
        - "find files with budget"
        - "email the meeting attendees"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
      responses:
        '200':
          description: Command processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '401':
          description: Not authenticated
        '500':
          description: Server error

  /api/v1/session:
    get:
      tags:
        - Session
      summary: Get session information
      description: Retrieve current session details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '404':
          description: No active session

  /api/v1/history:
    get:
      tags:
        - Session
      summary: Get command history
      description: Retrieve command history for current session
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of history items to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Command history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryItem'

  /api/v1/gmail/search:
    post:
      tags:
        - Gmail
      summary: Search Gmail
      description: Search emails using Gmail query syntax
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Gmail search query
                  example: "from:google is:unread"
                max_results:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  emails:
                    type: array
                    items:
                      type: object

  /api/v1/gmail/send:
    post:
      tags:
        - Gmail
      summary: Send email
      description: Send an email via Gmail
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - subject
                - body
              properties:
                to:
                  type: array
                  items:
                    type: string
                    format: email
                  example: ["user@example.com"]
                subject:
                  type: string
                  example: "Meeting Tomorrow"
                body:
                  type: string
                  example: "Let's meet at 2pm"
                cc:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: sent
                  message_id:
                    type: string

  /api/v1/calendar/events:
    get:
      tags:
        - Calendar
      summary: List calendar events
      description: Retrieve upcoming calendar events
      security:
        - BearerAuth: []
      parameters:
        - name: days_ahead
          in: query
          description: Number of days ahead to fetch
          schema:
            type: integer
            default: 7
            minimum: 1
            maximum: 365
        - name: max_results
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  events:
                    type: array
                    items:
                      type: object
    post:
      tags:
        - Calendar
      summary: Create calendar event
      description: Create a new calendar event
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - summary
                - start_time
              properties:
                summary:
                  type: string
                  example: "Team Meeting"
                start_time:
                  type: string
                  format: date-time
                  example: "2024-01-15T14:00:00"
                end_time:
                  type: string
                  format: date-time
                  example: "2024-01-15T15:00:00"
                description:
                  type: string
                  example: "Discuss Q4 planning"
                location:
                  type: string
                  example: "Conference Room A"
                attendees:
                  type: array
                  items:
                    type: string
                    format: email
                  example: ["user@example.com"]
      responses:
        '200':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: created
                  event_id:
                    type: string
                  event:
                    type: object

  /api/v1/drive/files:
    get:
      tags:
        - Drive
      summary: Search Drive files
      description: Search for files in Google Drive
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          description: Search query
          schema:
            type: string
        - name: mime_type
          in: query
          description: MIME type filter
          schema:
            type: string
            example: "application/pdf"
        - name: max_results
          in: query
          description: Maximum number of files to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  files:
                    type: array
                    items:
                      type: object

  /api/v1/drive/files/share:
    post:
      tags:
        - Drive
      summary: Share Drive file
      description: Share a file with another user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_id
                - email
              properties:
                file_id:
                  type: string
                  example: "1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                role:
                  type: string
                  enum: [reader, writer, commenter]
                  default: reader
      responses:
        '200':
          description: File shared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: shared
                  file_id:
                    type: string

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint

  schemas:
    CommandRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          description: Natural language command
          example: "search for emails from google"
        dry_run:
          type: boolean
          default: false
          description: Simulate without executing
        auto_confirm:
          type: boolean
          default: false
          description: Skip confirmation prompts

    CommandResponse:
      type: object
      properties:
        task_id:
          type: string
          nullable: true
          description: Task ID for async operations
        status:
          type: string
          enum: [queued, processing, completed, failed]
          description: Command status
        service:
          type: string
          nullable: true
          enum: [gmail, calendar, drive]
          description: Service used
        intent:
          type: string
          nullable: true
          description: Detected intent
        parameters:
          type: object
          nullable: true
          description: Extracted parameters
        result:
          description: Command result
          nullable: true
        error:
          type: string
          nullable: true
          description: Error message if failed
        timestamp:
          type: string
          format: date-time
          description: Command timestamp

    SessionInfo:
      type: object
      properties:
        session_id:
          type: string
          description: Session identifier
        started_at:
          type: string
          format: date-time
          description: Session start time
        command_count:
          type: integer
          description: Number of commands in session
        last_command:
          type: string
          nullable: true
          description: Last command executed

    HistoryItem:
      type: object
      properties:
        command:
          type: string
          description: Command text
        timestamp:
          type: string
          format: date-time
          description: Command timestamp
        service:
          type: string
          enum: [gmail, calendar, drive]
          description: Service used
        intent:
          type: string
          description: Detected intent
        success:
          type: boolean
          description: Whether command succeeded
        error:
          type: string
          nullable: true
          description: Error message if failed